requirement.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
decorator==5.1.1
Django==4.1.7
django-ckeditor==6.5.1
django-embed-video==1.4.8
django-js-asset==2.0.0
django-mathfilters==1.0.0
django-scrypt==0.2.3
djangorestframework==3.14.0
executing==1.2.0
feedparser==6.0.10
idna==3.4
install==1.3.5
ipython==8.11.0
jedi==0.18.2
jsonschema==4.17.3
matplotlib-inline==0.1.6
mysql-connector-python==8.0.32
numpy==1.24.2
pandas==2.0.0
parso==0.8.3
pdfkit==1.0.0
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.38
pure-eval==0.2.2
pycparser==2.21
Pygments==2.14.0
pyproject-toml==0.0.10
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
regex==2023.3.23
requests==2.28.2
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.2
toml==0.10.2
traitlets==5.9.0
tzdata==2022.7
urllib3==1.26.15
wcwidth==0.2.6
